requirements.txt
yfinance>=0.2.40
pandas>=2.0
pyyaml>=6.0
pytz>=2024.1
numpy>=1.24
